Minor fixes

+++ examples/single_file/example.py
@@ -96,9 +96,10 @@ def worker():
         open(f"{state_dir}/state_{args.current_host}", "wt").write("state")
         # write to the model output directory
         for file in Path(args.input_data).rglob("*"):
-            relp = file.relative_to(args.input_data)
-            path = Path(args.model_dir) / (str(relp) + "_proc_by_" + args.current_host)
-            path.write_text(file.read_text() + " processed by " + args.current_host)
+            if file.is_file():
+                relp = file.relative_to(args.input_data)
+                path = Path(args.model_dir) / (str(relp) + "_proc_by_" + args.current_host)
+                path.write_text(file.read_text() + " processed by " + args.current_host)
         open(f"{args.model_dir}/output_{args.current_host}", "wt").write("output")
     elif int(args.hps["task"]) == 2:
         logger.info(f"Input task2_data: {list(Path(args.input_task2_data).rglob('*'))}")
